Make internal builder class constructor private
