+++ larky/src/main/java/com/verygood/security/larky/modules/CodecsModule.java
@@ -3,7 +3,7 @@ package com.verygood.security.larky.modules;
 import com.google.common.primitives.Bytes;
 
 import com.verygood.security.larky.modules.codecs.TextUtil;
-import com.verygood.security.larky.modules.types.LarkyPByte;
+import com.verygood.security.larky.modules.types.LarkyByte;
 
 import net.starlark.java.annot.Param;
 import net.starlark.java.annot.ParamType;
@@ -81,7 +81,7 @@ public class CodecsModule implements StarlarkValue {
           @Param(
               name = "obj",
               allowedTypes = {
-                  @ParamType(type = LarkyPByte.class),
+                  @ParamType(type = LarkyByte.class),
               }
           ),
           @Param(
@@ -102,7 +102,7 @@ public class CodecsModule implements StarlarkValue {
           )
       }
   )
-  public String decode(LarkyPByte bytesToDecode, String encoding, String errors) {
+  public String decode(LarkyByte bytesToDecode, String encoding, String errors) {
       return TextUtil.starlarkDecodeUtf8(bytesToDecode.getBytes());
   }
 }
